Fix CodeQL alerts and formatting
